% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{searchLianjia}
\alias{searchLianjia}
\title{Scrape house price from the Lianjia website}
\usage{
searchLianjia(city = "sh", page_num = 2)
}
\arguments{
\item{city}{'sh' (shanghai)}

\item{page_num}{Number of pages to scrape}
}
\value{
A data frame
}
\description{
Scrape house price from the Lianjia website
}
\examples{
# In this version, we only support Shanghai city
price <- searchLianjia(city = 'sh', page_num = 100)
}
